# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DeutschLingoDeck Backend API
  version: 1.0.0
  description: REST API for DeutschLingoDeck, a German vocabulary learning platform.
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /dictionaries:
    get:
      tags: [Dictionaries]
      summary: List dictionaries owned by the current user
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: User dictionaries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedDictionarySummaryResponse"
...
  /statistics:
    get:
      tags: [Statistics]
      summary: Retrieve user dashboard statistics
      parameters:
        - $ref: "#/components/parameters/FromDate"
        - $ref: "#/components/parameters/ToDate"
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStatisticsResponse"
  /statistics/cards:
    get:
      tags: [Statistics]
      summary: Retrieve card statistics
      parameters:
        - name: dictionaryId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Card statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedCardStatisticsResponse"
...
  /statistics/history:
    get:
      tags: [Statistics]
      summary: Retrieve learning history
      parameters:
        - $ref: "#/components/parameters/FromDate"
        - $ref: "#/components/parameters/ToDate"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Learning history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedLearningHistoryResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    FromDate:
      name: from
      in: query
      required: false
      schema:
        type: string
        format: date
    ToDate:
      name: to
      in: query
      required: false
      schema:
        type: string
        format: date
...
  schemas:
...
    DictionarySummaryResponse:
      type: object
      required: [id, name, sourceLanguage, targetLanguage, cardCount, createdAt]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        sourceLanguage:
          type: string
        targetLanguage:
          type: string
        cardCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    DashboardStatisticsResponse:
      type: object
      properties:
        totalGames:
          type: integer
        totalAnswers:
          type: integer
        totalCorrectAnswers:
          type: integer
        totalIncorrectAnswers:
          type: integer
        overallAccuracy:
          type: number
          format: double
        totalStudyTimeSeconds:
          type: integer
          format: int64
        currentLearningStreak:
          type: integer
        longestLearningStreak:
          type: integer
    CardStatisticsResponse:
      type: object
      properties:
        cardId:
          type: integer
          format: int64
        dictionaryId:
          type: integer
          format: int64
        sourceText:
          type: string
        timesShown:
          type: integer
        timesCorrect:
          type: integer
        timesIncorrect:
          type: integer
        successRate:
          type: number
          format: double
        lastShownAt:
          type: string
          format: date-time
          nullable: true
        lastCorrectAt:
          type: string
          format: date-time
          nullable: true
...
    LearningHistoryResponse:
      type: object
      properties:
        gameId:
          type: integer
          format: int64
        dictionaryId:
          type: integer
          format: int64
        dictionaryName:
          type: string
        status:
          $ref: "#/components/schemas/GameStatus"
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
        accuracy:
          type: number
          format: double
    PagedDictionarySummaryResponse:
      allOf:
        - $ref: "#/components/schemas/PageResponse"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/DictionarySummaryResponse"
...
    PagedCardStatisticsResponse:
      allOf:
        - $ref: "#/components/schemas/PageResponse"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/CardStatisticsResponse"
    PagedLearningHistoryResponse:
      allOf:
        - $ref: "#/components/schemas/PageResponse"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/LearningHistoryResponse"
    PageResponse:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
...
    GameStatus:
      type: string
      enum:
        - CREATED
        - STARTED
        - IN_PROGRESS
        - FINISHED
        - ABANDONED
...